Crate kvdb

Key-Value store abstraction.

Structs

DBTransaction

Write transaction. Batches a sequence of put/delete operations for efficiency.

IoStats

Statistic for the span period

Enums

DBOp

Database operation.

IoStatsKind

Statistic kind to query.

Constants

PREFIX_LEN

Required length of prefixes.

Traits

KeyValueDB

Generic key-value database.

Functions

end_prefix

For a given start prefix (inclusive), returns the correct end prefix (non-inclusive). This assumes the key bytes are ordered in lexicographical order. Since key length is not limited, for some case we return None because there is no bounded limit (every keys in the serie [], [255], [255, 255] …).

Type Definitions

DBKey

Database keys.

DBValue

Database value.